refactor(mcp): move PlugWithConfig to its own file
Move `PtcRunnerMcp.Http.Server.PlugWithConfig` from server.ex to
`mcp_server/lib/ptc_runner_mcp/http/plug_with_config.ex` as
`PtcRunnerMcp.Http.PlugWithConfig`, per CLAUDE.md "Never nest multiple
modules in one file."
