+++ mcp_server/lib/ptc_runner_mcp/http/server.ex
@@ -1,11 +1,11 @@
 defmodule PtcRunnerMcp.Http.Server do
   @moduledoc false
 
-  alias PtcRunnerMcp.Http.Router
+  alias PtcRunnerMcp.Http.PlugWithConfig
 
   @spec child_spec(map()) :: {module(), keyword()}
   def child_spec(config) do
-    plug = {__MODULE__.PlugWithConfig, config}
+    plug = {PlugWithConfig, config}
 
     {Bandit,
      plug: plug,
@@ -19,18 +19,6 @@ defmodule PtcRunnerMcp.Http.Server do
     case :inet.parse_address(to_charlist(host)) do
       {:ok, ip} -> ip
       _ -> {127, 0, 0, 1}
-    end
-  end
-
-  defmodule PlugWithConfig do
-    @moduledoc false
-
-    def init(config), do: config
-
-    def call(conn, config) do
-      conn
-      |> Plug.Conn.put_private(:ptc_http_config, config)
-      |> Router.call([])
     end
   end
 end
